requirements.txt
aiogram>2
PyYAML>5
python-dateutil~=2.8.1